requirements.txt
fastapi-0.115.3
requests
uvicorn-0.32.0